@@ -15,72 +15,72 @@
     "publish-@candiman/website": "./publish-@candifood-website.sh",
     "gcloud-deploy-no-promot": "gcloud app deploy --no-promote",
     "gcloud-deploy": "gcloud app deploy",
-    "bundle-report": "ng build --prod --stats-json && webpack-bundle-analyzer dist/candifood-ui/stats.json"
+    "bundle-report": "ng build --prod --stats-json && webpack-bundle-analyzer dist/candifood-ui/stats.json",
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
   },
   "private": true,
   "dependencies": {
     "@agm/core": "1.1.0",
-    "@angular/animations": "8.2.14",
+    "@angular/animations": "9.0.0",
     "@angular/cdk": "8.0.0",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/elements": "8.2.14",
-    "@angular/forms": "8.2.14",
+    "@angular/common": "9.0.0",
+    "@angular/compiler": "9.0.0",
+    "@angular/core": "9.0.0",
+    "@angular/elements": "9.0.0",
+    "@angular/forms": "9.0.0",
     "@angular/material": "8.0.0",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/platform-browser": "9.0.0",
+    "@angular/platform-browser-dynamic": "9.0.0",
+    "@angular/router": "9.0.0",
     "@fortawesome/angular-fontawesome": "0.5.0",
     "@fortawesome/fontawesome-svg-core": "1.2.25",
     "@fortawesome/free-brands-svg-icons": "5.11.2",
     "@fortawesome/free-regular-svg-icons": "5.11.2",
     "@fortawesome/free-solid-svg-icons": "5.11.2",
-    "@nebular/auth": "4.4.0",
-    "@nebular/security": "4.4.0",
-    "@nebular/theme": "4.4.0",
-    "@ng-bootstrap/ng-bootstrap": "5.1.2",
-    "@webcomponents/custom-elements": "1.3.0",
-    "angularx-social-login": "2.1.1",
-    "bootstrap": "4.3.1",
-    "core-js": "3.1.3",
+    "@nebular/auth": "4.6.0",
+    "@nebular/security": "4.6.0",
+    "@nebular/theme": "4.6.0",
+    "@ng-bootstrap/ng-bootstrap": "5.2.1",
+    "@webcomponents/custom-elements": "1.3.2",
+    "angularx-social-login": "2.2.1",
+    "bootstrap": "4.4.1",
+    "core-js": "3.6.4",
     "document-register-element": "1.14.3",
     "hammerjs": "2.0.8",
     "jquery": "3.4.1",
-    "ngx-infinite-scroll": "7.2.0",
+    "ngx-infinite-scroll": "8.0.1",
     "rxjs": "6.5.4",
-    "tslib": "1.9.3",
-    "zone.js": "0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.25",
-    "@angular-devkit/build-ng-packagr": "0.803.25",
-    "@angular/cli": "8.3.25",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "~0.900.1",
+    "@angular-devkit/build-ng-packagr": "~0.900.1",
+    "@angular/cli": "9.0.1",
+    "@angular/compiler-cli": "9.0.0",
+    "@angular/language-service": "9.0.0",
     "@schematics/angular": "8.0.0",
     "@types/gapi": "0.0.36",
     "@types/gapi.auth2": "0.0.50",
-    "@types/jasmine": "2.8.12",
-    "@types/jasminewd2": "2.0.6",
-    "@types/node": "10.12.24",
+    "@types/jasmine": "3.5.3",
+    "@types/jasminewd2": "2.0.8",
+    "@types/node": "^13.7.0",
     "chalk": "2.4.2",
-    "codelyzer": "5.2.1",
+    "codelyzer": "^5.1.2",
     "gulp": "4.0.2",
-    "jasmine-core": "3.3.0",
+    "jasmine-core": "3.5.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.1.0",
-    "karma-chrome-launcher": "2.2.0",
+    "karma": "4.4.1",
+    "karma-chrome-launcher": "3.1.0",
     "karma-coverage-istanbul-reporter": "2.0.4",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.0",
-    "ng-packagr": "5.7.1",
-    "ngx-build-plus": "8.1.5",
+    "ng-packagr": "^9.0.0",
+    "ngx-build-plus": "9.0.6",
     "protractor": "5.4.2",
-    "ts-node": "8.2.0",
-    "tslib": "1.9.3",
-    "tslint": "5.17.0",
-    "typescript": "3.5.3",
-    "webpack-bundle-analyzer": "3.3.2"
+    "ts-node": "8.6.2",
+    "tslint": "6.0.0",
+    "typescript": "3.7.5",
+    "webpack-bundle-analyzer": "3.6.0"
   }
 }
